Break example into multiple files
